re-export enough stuff to de-break the non-doctest tests
