+++ tokio-trace/Cargo.toml
@@ -15,4 +15,5 @@ ansi_term = "0.11"
 humantime = "1.1.1"
 env_logger = "0.5"
 tokio-trace-log = { path = "../tokio-trace-log" }
+tokio-trace-core = { path = "../tokio-trace-core", features = ["test-support"] }
 futures = "0.1"
